Document `package.json` entry `"flat": true` (#162)
This documents the ability to add `"flat": true` as an entry in `package.json`
to enforce the semantics of `yarn install --flat` on every installation, as
implemented in #259. This also notes the drawbacks involved in doing so
(forcing flatness upon consumers of the package).

**This is not ready to merge yet because:**
The only outstanding issue is the link generated using `{{url_base}}` doesn't
include the language, despite it working on other pages. Let me know how to fix
that up and I'll take care of it :)

Also open to rewording things.