Update docs for `add` to clarify that it can work with multiple packages. (#228)
The docs were explicitly refering to acting on a single package, which was inaccurate.

The ellipse in `<package...>` is there to imply that multiple packages
are supported, but the notation is universally recognized. Updating
the text improves the clarity that the `add` sub-commnand can work on
multiple dependencies at once.