Packages search: allow searching for packages on every pages (#383)
* chore(tooling): Add webpack and babel to allow ES6 code

The JavaScript code of the website itself starts to be bigger than in
the beginings. To allow for a greater dev experience (write modern
JavaScript) and plan next iterations on the search and package detail
pages (use React), this PR propose to add Babel and Webpack to the
toolbet of the website.

The structure of the JS files now is (in my opinion) more reasonable to think about,
and also will allow for faster development for further features.

* fix non-display of initial results on Safari

fixes #4

Safari is more strict about closing brackets in querySelectors than Chrome, which just closes it for you

* use hashes in webpack build

* initial add of preact

* add common.js always

* order of operations

* fix typo and add polyfill

* fix typo

* make webpack work slightly faster on development

* put very basic instantsearch-react in there

doesn't work with preact for some reason. Will troubleshoot later

* some jsx filling in

* move html structure

* search inside different file

* initialising instant search

* use classname; clean

* add icons

* put common.js manually

ensure it doesn't get overwritten

* move order of operations

* not every item has a license

* move Search to folder to be able to split up

* move to separate files per component

* extend bootstrap in our pagination

we can't really add custom classes to the algolia pagination

* cleanup

* add possible faceting

* add placeholder mixin

* put placeholder

* style searchbox

* border images

* add container around results

* hide comment

* switch out moment for date-fns

* make the hero not float down on small sizes

* center pagination

* update react-instantsearch

* make <main> and reset appear/disappear

* don't do the redirect yet

* don't link to /packages yet

* no packages page yet

* no dedicated search page

the /packages will later be occupied by the detail page

* add very light to the colors

* wip: make active color darker

* active background

* add homepage

* update dependencies

* make urls work

* in 3.2.0 key for configure became not necessary

bug/feature was fixed in react-instantsearch

* wip of keywords

* don't show faceting

* remove preact alias (didn't work)

* only update location when needed, otherwise it makes the UI slow

* save newPush when it's pushed

this will fix the updating while you're typing fast, but doesn't catch it if it's the last character typed

* keywords doesn't matter

* no connectedToggle

issue is still open with nested widgets

* don't show refinementlist

* add packages page

same as what used to be done in /search, but now in React

* make search be on the packages url

* separate component for HasResults

* don't mess with the react state for things that don't change rendering

* chore(dev): webpack build speed optimization

* fix(keys): fix react keys

* fix(url): good url sync handling

* "static rendering" of search component

* no log

* add "key" to lists of components

* put back margin on hero

* don't override docsearch style

this didn't get applied on the version that's now live, but for some reason it does count now. It doesn't really work, so I'm going back to the defaults

* use prettier

settings:

* with single-quotes
* with trailing comma

* query -> q

* run prettier on webpack

* stub out the featured packages so they don't "jump in"

* put page in url

* add title to history

* no results "screen"

* make the cross icon take up place regardless of state

suggested by @cpojer in chat

* make sure input doesn't have border-radius

iOS is naughty and gives us radii all the time

* small css fixes

* chore(prettier): use precommit hook from prettier docs

* style(searchbar): be more precise about position

there seems to be an iOS Safari bug where the buttons don't show up on the right place. This should fix that

* chore(dev): add vendor, start i18n in js

* chore(pagesize): avoid babel-polyfill

since the only place this is needed now is to iterate of this NodeList, we can replace it safely and save a considerable abount of space

* chore(build): fix production build

* chore(build): fix production build

* bootstrap fix

* build optimization

* build optimization

* remove bundle analyze

* remove lodash optimizations, failing

* feate(i18n): use i18n values for searchbar cue

* chore(bundle): include bootstrap collapse

this is needed for the 🍔menu

* nit(search): don't put a default in the code for placeholder

* feat(searchbox): use own searchbox component

the way to go for react-instantsearch is to make sure that our widgets are reusable and "copy-paste"able. This is an experiment of how this is done.

The reason a custom SearchBox is necessary is because we need events to add a class around the root when it's focused

* style(searchbox): style ":active"

things needed for this:

1. different way of referencing the border-images
2. custom searchbox
3. other image

* fix(search): add pointer cursors to buttons

* fix(explore): list /packages under language

it should be fully i19d

* feat(search): internationalise search by algolia

* fix(search): url is internationalised too

* fix(search): same margin left and right

* fix(search): only show keywords when not empty

* feat(search): no pagination when there's only one page

* fix(searchbar): color with mix instead of darken()

cc @lukyvj

* feat(search): make "search" show up in iOS keyboard

needs a form with action and an inputtype=search with a name

* fix(search): don't autocorrect etc. input

this usually causes you to type a certain package name and then it's autocorrected into something irrelevant

* feat(search): blur input on submit

this causes the keyboard to disappear on mobile leaving you more place to explore the results

* style(search): make input noticeable

use a white background, which is better for a11y, as suggested by @thejameskyle

* style(search): give searchbar a bit more margin

suggested by @lukyvj

* fix(build): run webpack before jekyll

Jekyll doesn't rebuild the files when the `webpack.json` doesn't exist yet

* fix(pagesize): import directly from react-instantsearch core

this is to save size in the final build

* fix(search): put more exact min-height to wrapper to avoid jump

this is only relevant on empty cache where the js isn't loaded yet

* fix(i18n): json encode translation string

this is the case because some languages or strings could contain ' and break the string.