feat(deprecation): show deprecated message (#520)
The `deprecated` value in the index is either `false` or a string. This adds that string as a `title` to that badge. In the future this message would be visible more clearly, but I couldn't think of a good UI to put it, so this is a nice inbetween state where you can already read the message if you want

related: https://github.com/algolia/yarnpkg-website/issues/20#issuecomment-304726188